pyproject.toml
[project]
name = "power-bi-regression-testing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "certifi==2025.7.14",
    "cffi==1.17.1",
    "charset-normalizer==3.4.2",
    "clr-loader==0.2.7.post0",
    "cryptography==45.0.5",
    "idna==3.10",
    "msal==1.33.0",
    "numpy==2.3.2",
    "pandas==2.3.1",
    "pyarrow==21.0.0",
    "pycparser==2.22",
    "pyjwt==2.10.1",
    "python-dateutil==2.9.0.post0",
    "pythonnet==3.0.5",
    "pytz==2025.2",
    "pywin32==311",
    "requests==2.32.4",
    "six==1.17.0",
    "tabulate==0.9.0",
    "tzdata==2025.2",
    "urllib3==2.5.0",
]
